news_api/admin_api/requirements.txt
Django>=4.2
djangorestframework
gunicorn
psycopg2-binary
django-redis
django-filter
drf-yasg
python-decouple
